refactor: remove sub-package
